fix seeking of journal-cursor upon startup
due to strange behaviour in sd_journal_next(), some random starting point in
the journals were used instead of the specified "tail". as a result, lots of
log messages were (potentially) resend to the graylog server.

to nudge the sd_journal_next() call into the right behaviour, a workaround is
used, similar to what's outlined in the various related bugreports:

* https://bugzilla.redhat.com/show_bug.cgi?id=979487
* https://bugzilla.freedesktop.org/show_bug.cgi?id=64614
* https://lists.freedesktop.org/archives/systemd-devel/2014-October/023923.html